Add notes to README.md
The process followed in the existing code utilizes the mime types from
the /etc folder. It contains a large number of extensions and that makes
the process of generating a dircolors file long and boring.

I could utilize the dircolors --print-database command to generate a
shorter version with less extensions. I must still review for potential
conflicts and issues.